move packets to their own database
