+++ src/commonMain/kotlin/io/rebble/libpebblecommon/main.kt
@@ -1,4 +1,6 @@
 package io.rebble.libpebblecommon
 
+import io.rebble.libpebblecommon.packets.PhoneAppVersion
+
 @OptIn(ExperimentalUnsignedTypes::class)
 expect fun getPlatform(): PhoneAppVersion.OSType
